examples/agents/agent_with_fargate_tool/cdk/requirements.txt
aws-cdk-lib==2.196.0
boto3~=1.34.103
cdk_nag==2.28.196
